Update OrderProduct model to set price before create and calculate subtotal

+++ db/migrate/20240715010513_add_price_to_order_products.rb
@@ -1,5 +1,5 @@
 class AddPriceToOrderProducts < ActiveRecord::Migration[7.1]
   def change
-    add_column :order_products, :price, :decimal
+    add_column :order_products, :price, :integer
   end
 end
